git-svn-id: svn://svn.valgrind.org/valgrind/trunk@8369

+++ coregrind/m_debuginfo/tytypes.c
@@ -160,7 +160,7 @@ static void pp_XArray_of_pointersOrRefs ( XArray* xa ) {
    VG_(printf)("{");
    for (i = 0; i < VG_(sizeXA)(xa); i++) {
       void* ptr = *(void**) VG_(indexXA)(xa, i);
-      VG_(printf)("0x%05lx", ptr);
+      VG_(printf)("0x%05lx", (unsigned long)(ptr));
       if (i+1 < VG_(sizeXA)(xa))
          VG_(printf)(",");
    }
@@ -174,7 +174,7 @@ void ML_(pp_D3Expr) ( D3Expr* expr ) {
 }
 void ML_(pp_TyField) ( TyField* field ) {
    VG_(printf)("TyField(0x%05lx,%p,\"%s\")",
-               field->typeR, field->loc,
+               (unsigned long)(field->typeR), field->loc,
                field->name ? field->name : (UChar*)"");
 }
 void ML_(pp_TyBounds) ( TyBounds* bounds ) {
@@ -221,9 +221,9 @@ void ML_(pp_Type) ( Type* ty )
          break;
       case Ty_PorR:
          VG_(printf)("Ty_PorR(%d,%c,0x%05lx)",
-                     ty->Ty.PorR.szB, 
+                     ty->Ty.PorR.szB,
                      ty->Ty.PorR.isPtr ? 'P' : 'R',
-                     ty->Ty.PorR.typeR);
+                     (unsigned long)(ty->Ty.PorR.typeR));
          break;
       case Ty_Enum:
          VG_(printf)("Ty_Enum(%d,%p,\"%s\")",
@@ -235,7 +235,7 @@ void ML_(pp_Type) ( Type* ty )
          break;
       case Ty_StOrUn:
          if (ty->Ty.StOrUn.complete) {
-            VG_(printf)("Ty_StOrUn(%d,%c,%p,\"%s\")",
+            VG_(printf)("Ty_StOrUn(%ld,%c,%p,\"%s\")",
                         ty->Ty.StOrUn.szB, 
                         ty->Ty.StOrUn.isStruct ? 'S' : 'U',
                         ty->Ty.StOrUn.fields,
@@ -250,13 +250,13 @@ void ML_(pp_Type) ( Type* ty )
          break;
       case Ty_Array:
          VG_(printf)("Ty_Array(0x%05lx,%p)",
-                     ty->Ty.Array.typeR, ty->Ty.Array.bounds);
+                     (unsigned long)(ty->Ty.Array.typeR), ty->Ty.Array.bounds);
          if (ty->Ty.Array.bounds)
             pp_XArray_of_pointersOrRefs( ty->Ty.Array.bounds );
          break;
       case Ty_TyDef:
          VG_(printf)("Ty_TyDef(0x%05lx,\"%s\")",
-                     ty->Ty.TyDef.typeR,
+                     (unsigned long)(ty->Ty.TyDef.typeR),
                      ty->Ty.TyDef.name ? ty->Ty.TyDef.name
                                          : (UChar*)"" );
          break;
@@ -265,7 +265,7 @@ void ML_(pp_Type) ( Type* ty )
          break;
       case Ty_Qual:
          VG_(printf)("Ty_Qual(%c,0x%05lx)", ty->Ty.Qual.qual,
-                     ty->Ty.Qual.typeR);
+                     (unsigned long)(ty->Ty.Qual.typeR));
          break;
       case Ty_Void:
          VG_(printf)("Ty_Void%s",
